`PrecisionRecallF1ScoreMeter` for Non-1D Arrays and Removing the Assertion for num_classes>1 (#491)
* Fixing catalyst-team#488 by changing the assert statements in PrecisionRecallF1ScoreMeter.add()

Just removed the assert statements that caused the "bug" (it wasn't needed). The current implementation now works with any shape of array because np.sum reduces the arrays to 1 value as the sum. Note: Also, added a check for the number of elements to be equal.

* Added a test with dummy seg tensors for PrecisionrecallF1ScoreMeter

For changes in commit feafdbd (previous commit)

* Removed the assert statement (num_classes > 1)

Did so because it makes sense to have num_classes=1 when you have binary labels that are single channel and predictions are single channel

* Removed the batch size check because the number of elements check is the most important one

Batch size isn't necessary for .add() to work properly, just recommended

* Refactored tests and updated them so that PrecisionRecallF1Meter works with any shaped array

Did so because batch size isn't necessary, just recommended

* Fixing the name of `tests_on_meter_counts_and_value` to `runs_tests_on_meter_counts_and_value` to not conflict with pytest