Addressing #638 to fix the MulticlassDiceMetricCallback (WIP) (#643)
* Added the MulticlassDiceMetricCallback to the __init__.py for callbacks so it actually works when importing from catalyst.callbacks

* Addresses #638, removes unused arguments, and adds documentation

Made sure to keep class_names functionality and refactored the code to be a bit more readable

* Code style fixes

* Fixing import order

* Now averaging only the classes specified in `class_names` when `class_names` is not None

Doing so because before, you would specify certain classes in `class_names` and it would return the mean of all of the classes. This is confusing and misleading, so I changed it to the new way.